Treewide: Move to databinding wherever possible
N.B. I realized that the shared viewmodel is getting bloated
with this changes, optimally we probably should separate it
like we used to. In hindsight the shared model should trim
the code but alas it's not. For now, I leave it as it is
as I don't want to refactor this again.